- fixed annotated text shown in the summary window
when an  annotated buffer (that is  visiting a file) is  modified in a
way that changes the position of an annotation in that buffer, showing
the  summary window  -without saving  the  file first-  will read  the
annotation using the updated position but getting the content from the
outdated file; this  gives wrong results in the  text annotation parts
of  the  summary windows,  with  the  text  shifted from  the  correct
position.

This patch try to fix this bug.